# docs/swagger.yaml
basePath: /api/v1
definitions:
  helpers.HTTPError:
    properties:
      code:
        type: integer
      error:
        type: string
    type: object
...
  models.ResponseMessage:
    properties:
      message:
        example: runner successfully deleted
        type: string
    type: object
...
  models.Task:
    properties:
      command:
        type: string
      name:
        type: string
      runner:
        type: string
      schema:
        additionalProperties: {}
        type: object
      synchronous:
        type: boolean
    required:
    - command
    - name
    - runner
    type: object
...
info:
  contact:
    email: info@kubecode.io
    name: Kubecode Support
    url: https://www.kubecode.io/contact
  description: API Gateway for your kubernetes services.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Kriten
  version: v0.3
paths:
...
  /tasks:
    get:
      consumes:
      - application/json
      description: List all tasks available on the cluster
      parameters:
      - description: Maximum number of tasks to return (default 100)
        in: query
        name: limit
        type: integer
      - description: Number of tasks to skip (default 0)
        in: query
        name: offset
        type: integer
      - description: Filter by task name
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Task'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helpers.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.HTTPError'
      security:
      - Bearer: []
      summary: List all tasks
      tags:
      - tasks
    post:
      consumes:
      - application/json
      description: Add a task to the cluster
      parameters:
      - description: New task
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/models.Task'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Task'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helpers.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.HTTPError'
      security:
      - Bearer: []
      summary: Create a new task
      tags:
      - tasks
  /tasks/{name}:
    delete:
      consumes:
      - application/json
      description: Delete by task name
      parameters:
      - description: Task name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/models.ResponseMessage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helpers.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.HTTPError'
      security:
      - Bearer: []
      summary: Delete a task
      tags:
      - tasks
    get:
      consumes:
      - application/json
      description: Get information about a specific task
      parameters:
      - description: Task name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Task'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helpers.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.HTTPError'
      security:
      - Bearer: []
      summary: Get a task
      tags:
      - tasks
    patch:
      consumes:
      - application/json
      description: Update a task in the cluster
      parameters:
      - description: Task name
        in: path
        name: name
        required: true
        type: string
      - description: Update task
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/models.Task'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Task'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helpers.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.HTTPError'
      security:
      - Bearer: []
      summary: Update a task
      tags:
      - tasks
  /tasks/{name}/schema:
    delete:
      consumes:
      - application/json
      description: Remove validation schema associated to a specific task
      parameters:
      - description: Task name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helpers.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.HTTPError'
      security:
      - Bearer: []
      summary: Delete schema
      tags:
      - tasks
    get:
      consumes:
      - application/json
      description: Get validation schema associated to a specific task
      parameters:
      - description: Task name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helpers.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.HTTPError'
      security:
      - Bearer: []
      summary: Get schema
      tags:
      - tasks
    post:
      consumes:
      - application/json
      description: Add or Update validation schema associated to a specific task
      parameters:
      - description: Task name
        in: path
        name: name
        required: true
        type: string
      - description: New schema
        in: body
        name: schema
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helpers.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.HTTPError'
      security:
      - Bearer: []
      summary: Update schema
      tags:
      - tasks
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"